Use wrapped Writer interface, refactor global cursor helpers to utils
