+++ utils.go
@@ -1,8 +1,74 @@
 package cursor
 
-import "io"
+import (
+	"io"
+	"os"
+)
+
+//
+// Helpers for global cursor handling on os.Stdout
+//
 
 var autoheight int
+var cursor = &Cursor{writer: os.Stdout}
+
+func SetTarget(w Writer) {
+	cursor = cursor.WithWriter(w)
+}
+
+// Up moves the cursor n lines up relative to the current position.
+func Up(n int) {
+	cursor.Up(n)
+	autoheight += n
+}
+
+// Down moves the cursor n lines down relative to the current position.
+func Down(n int) {
+	cursor.Down(n)
+	if autoheight > 0 {
+		autoheight -= n
+	}
+}
+
+// Right moves the cursor n characters to the right relative to the current position.
+func Right(n int) {
+	cursor.Right(n)
+}
+
+// Left moves the cursor n characters to the left relative to the current position.
+func Left(n int) {
+	cursor.Left(n)
+}
+
+// HorizontalAbsolute moves the cursor to n horizontally.
+// The position n is absolute to the start of the line.
+func HorizontalAbsolute(n int) {
+	cursor.HorizontalAbsolute(n)
+}
+
+// Show the cursor if it was hidden previously.
+// Don't forget to show the cursor at least at the end of your application.
+// Otherwise the user might have a terminal with a permanently hidden cursor, until they reopen the terminal.
+func Show() {
+	cursor.Show()
+}
+
+// Hide the cursor.
+// Don't forget to show the cursor at least at the end of your application with Show.
+// Otherwise the user might have a terminal with a permanently hidden cursor, until they reopen the terminal.
+func Hide() {
+	cursor.Hide()
+}
+
+// ClearLine clears the current line and moves the cursor to it's start position.
+func ClearLine() {
+	cursor.ClearLine()
+}
+
+// Clear clears the current position and moves the cursor to the left.
+func Clear() {
+	cursor.Clear()
+}
 
 // Bottom moves the cursor to the bottom of the terminal.
 // This is done by calculating how many lines were moved by Up and Down.
